Test Emacs 26.0.91 on Travis
This required updating a few tests for elisp checkers, and two workarounds.

The first workaround is to install GnuTLS 3 on CI.  Without it, Emacs 26 fails
to fetch from ELPA.  I did not find binaries for GnuTLS, so we have to compile
it from source.  We only have to do it for Emacs 26 and snapshot though.

The second workaround is to add `--verbose` to `cask install`.  See
cask/cask#367.  This is harmless for all Emacs versions.

On a related note, llvm-mode has been removed from MELPA because the LLVM git
repository is too large.  We'll have to fetch the mode directly if we want to
run all of our integration tests on CI.

See melpa/melpa/commits/aa2d7a88731ecd9410ab7dc887a7a5f837d8a93b

The JSON parser has become stricter in Emacs 26, and does not allow newlines or
tabs inside strings.  I fixed the test data that had those invalid characters.

The check-declare output format has changed.  The column is now optional.  This
required yet another version guard in the test.  Not sure this is not a
regression though.

Paths returned by check-declare are now relative to the output; the root file we
used in the example file ended up with a bunch of "../../…".  Removing the `/`
makes the test more robust.

For the emacs-lisp checker tests, `message` can now be called without arguments.
so I've changed the tests to use another function.

Alos, the checker now emits a new error when compressed files do not match the
`provided` directive.  Using the correct name will make this test
retro-compatible.